Listen on Marathon event bus to initiate DNS updates
These changes switch from a purely polling mechanism to a hybrid
approach where the program listens for status updates related to
the app before refreshing state from the Marathon API, preventing
needless updates when the state has not changed.

The current logic isn't perfect and could be improved by debouncing
status update messages until a quiescence period is achieved, then
issuing a single update. This would prevent the repetitive updates
incurred when many tasks are moving around and changing state, but
this is most likely a minor issue except in the most extreme of
cases.